refactor (client): import/export savePhoto function
