+++ client/pages/platforms-all.html
@@ -44,6 +44,7 @@
 </body>
 <script src="../js/Authentication/redirect.js"></script>
 <script src="../js/Authentication/logout.js"></script>
-<script src="../js/fetchPhotos.js" type="module"></script>
-<script src="../js/carousel.js" type="module"></script>
+<script src="../js/fetchPhotos.js"></script>
+<script src="../js/savePhoto.js" type="module" defer></script>
+<script src="../js/carousel.js" type="module" defer></script>
 </html>
